pyproject.toml

[tool.poetry]
name = "nav7"
version = "1.0.0"
description = "[Na7] AI Agent career navigation project"
authors = ["Daeun <rla0576@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

langchain = "0.2.15"
langchain-core = "0.2.37"
langchain-openai = "0.1.23"
openai = "1.55.3"
chromadb = "0.4.24"
chroma-hnswlib = "0.7.3"
fastapi = "0.112.2"
uvicorn = "0.30.6"
python-multipart = "0.0.9"
pydantic = "2.9.0b2"
gitpython = "3.1.43"
python-dotenv = "1.0.1"
requests = "2.32.3"
aiohttp = "3.10.5"
tqdm = "4.66.5"
langchain-experimental = "0.0.61"
langchain-community = "0.2.15"
langchain-anthropic = "0.1.23"
langchain-text-splitters = "0.2.2"
langchain-elasticsearch = "0.2.2"
langchain-upstage = "0.1.8"
langchain-chroma = "0.1.3"
langchain-cohere = "0.1.9"
langchain-milvus = "0.1.4"
langchain-google-genai = "1.0.10"
langchainhub = "0.1.21"
langgraph = "0.2.14"
langsmith = "0.1.108"
deepl = "1.18.0"
kiwipiepy = "0.17.1"
konlpy = "0.6.0"
pandas = "2.2.2"
redis = "5.0.8"
jupyter = "1.1.1"
jupyterlab = "4.2.5"
ipykernel = "6.29.5"
tavily-python = "0.7.3"
httpx = "0.27.2"
sentence-transformers = "^4.1.0"
feedparser = "^6.0.11"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"